@@ -19,7 +19,7 @@ package org.secuso.privacyfriendlytodolist.view;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 
 /**
  * Created by yonjuni on 17.11.16.
