@@ -18,7 +18,7 @@
 package org.secuso.privacyfriendlytodolist.view;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.AttributeSet;
 import android.view.View;
 
