@@ -20,7 +20,7 @@ package org.secuso.privacyfriendlytodolist.view.dialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.preference.PreferenceManager;
-import android.support.v7.app.AlertDialog;
+import androidx.appcompat.app.AlertDialog;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
